backend/pyproject.toml
[tool.poetry]
name = "fastapi-ai-agent"
version = "0.1.0"
description = ""
authors = ["mazzasaverio <saverio3107@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.29"
loguru = "^0.7.2"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.16"
pyyaml = "^6.0.1"
langchain = "^0.1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
